float acceleration

+++ inc/sensors/accelerometer.h
@@ -9,6 +9,6 @@ short acc_power_on();
 
 /// @brief Возвращает ускорение по трем осям
 /// @param xyz массив из трех значений, в котором будет хранится ускорение по осям x,y,z соответственно
-void read_acceleration_xyz(double* buffer_xyz);
+void read_acceleration_xyz(float* buffer_xyz);
 
-void read_acceleration_angular_xyz(double* buffer_xyz);
+void read_acceleration_angular_xyz(float* buffer_xyz);
